Initialize the Tito Widget after everything else has loaded
This commit does two things:

1) Uses the JS API to mount the widget rather than the custom component
2) Waits for 1s before mounting the widget

The wait gives the rest of the page time to load before
loading the widget.

Better would be to do this on some kind of callback or event when
the rest of the page is still building.

OR figuring out what part of that is causing the page to be over-written
after the widget is mounted.

But this hack will probably work 99% of the time.